+++ docker/pytorch/pyproject.toml
@@ -18,9 +18,10 @@ dependencies = [
 [project.optional-dependencies]
 sagemaker = [
     "smclarify",
-    "sagemaker>=3.0.0",
+    "sagemaker>=3.4.0",
     "sagemaker-pytorch-training",
     "sagemaker-training",
+    "mlflow>=3.9.0",
     "bokeh",
     "imageio",
     "numba",
