+++ test/pytorch/integration/sagemaker/test_sm_training.py
@@ -1,0 +1,92 @@
+"""SageMaker training integration tests for PyTorch DLC.
+
+Uses SageMaker Python SDK v3 (ModelTrainer API).
+
+Minimal test set covering P0 functionalities:
+  F1: Distributed training with nccl (test_mnist_distributed_gpu)
+  F2: torch.distributed primitives (test_dist_operations_gpu)
+
+Tests launch real SageMaker training jobs — no GPU needed on the runner.
+"""
+
+import os
+
+from sagemaker.core.training.configs import Compute, InputData, SourceCode
+from sagemaker.train import ModelTrainer
+from sagemaker.train.distributed import Torchrun
+from test_utils import random_suffix_name
+
+RESOURCE_DIR = os.path.join(os.path.dirname(__file__), "resources")
+INSTANCE_TYPE = "ml.g4dn.12xlarge"
+IMAGE_URI = os.environ["TEST_IMAGE_URI"]
+MNIST_S3_URI = "s3://dlc-cicd-models/pytorch/sagemaker-test-data/"
+
+
+def _run_sm_training(
+    image_uri,
+    entry_script,
+    source_dir,
+    instance_type,
+    instance_count,
+    hyperparameters=None,
+    environment=None,
+    input_data=None,
+    job_name_prefix="pt-sm-test",
+):
+    """Launch a SageMaker training job using ModelTrainer (SDK v3) and wait for completion."""
+    source_code = SourceCode(
+        source_dir=source_dir,
+        entry_script=entry_script,
+    )
+
+    compute = Compute(
+        instance_type=instance_type,
+        instance_count=instance_count,
+    )
+
+    model_trainer = ModelTrainer(
+        training_image=image_uri,
+        source_code=source_code,
+        compute=compute,
+        role=os.environ.get("SM_ROLE_ARN"),
+        base_job_name=random_suffix_name(job_name_prefix, 32),
+        hyperparameters=hyperparameters or {},
+        environment=environment or {},
+        distributed=Torchrun() if instance_count > 1 else None,
+    )
+
+    model_trainer.train(input_data_config=input_data, wait=True)
+
+
+def test_mnist_distributed_gpu():
+    """F1: 2-node distributed GPU training with nccl backend."""
+    _run_sm_training(
+        image_uri=IMAGE_URI,
+        entry_script="mnist.py",
+        source_dir=RESOURCE_DIR,
+        instance_type=INSTANCE_TYPE,
+        instance_count=2,
+        hyperparameters={"backend": "nccl", "epochs": "1"},
+        environment={"FI_EFA_FORK_SAFE": "1"},
+        input_data=[
+            InputData(
+                channel_name="training",
+                data_source=MNIST_S3_URI,
+            )
+        ],
+        job_name_prefix="pt-mnist-nccl",
+    )
+
+
+def test_dist_operations_gpu():
+    """F2: torch.distributed primitives (all_reduce, broadcast, etc.) on GPU."""
+    _run_sm_training(
+        image_uri=IMAGE_URI,
+        entry_script="distributed_operations.py",
+        source_dir=RESOURCE_DIR,
+        instance_type=INSTANCE_TYPE,
+        instance_count=2,
+        hyperparameters={"backend": "nccl"},
+        environment={"FI_EFA_FORK_SAFE": "1"},
+        job_name_prefix="pt-dist-ops",
+    )
