--- test/pytorch/sagemaker/requirements.txt
@@ -1,3 +1,0 @@
-boto3
-pytest
-sagemaker>=3.0.0
